% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrale_utils.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot a Spectral Trace}
\usage{
plot_trace(spectra)
}
\arguments{
\item{spectra}{\link[data.table:data.table]{data.table} -- the return of \link{spectracle}.}
}
\value{
A plot is printed to the active device.
}
\description{
Plot a Spectral Trace
}
