% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectracle-package.R
\docType{package}
\name{spectracle-package}
\alias{spectracle-package}
\title{spectracle: Derive Spectra from Full Spectrum Cytometry Controls}
\description{
Fully automated derivation and characterization of spectra from full spectrum cytometry controls -- 'making a spectacle of your spectra'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nlaniewski/spectracle}
  \item Report bugs at \url{https://github.com/nlaniewski/spectracle/issues}
}

}
\author{
\strong{Maintainer}: Nathan Laniewski \email{nathan_laniewski@urmc.rochester.edu}

Authors:
\itemize{
  \item Oliver Burton \email{olivertburton@gmail.com} (\href{https://orcid.org/0000-0003-3884-7373}{ORCID})
}

}
\keyword{internal}
