% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectracle.R
\name{spectracle}
\alias{spectracle}
\title{Derive Spectra from Full Spectrum Cytometry Controls}
\usage{
spectracle(
  raw.reference.controls,
  filter.top.expressing = NULL,
  top.expressing.override = NULL,
  print.timings = FALSE,
  ...
)
}
\arguments{
\item{raw.reference.controls}{a character vector of full directory/path names to raw single-color control .fcs files; one of \link{list.dirs} or \link{list.files}.}

\item{filter.top.expressing}{a character vector -- default \code{NULL}; to modulate the automated selection of 'spectral.events', the supplied character vector should match to unique .fcs sample name(s). If defined, a conditional filtering will take place to return only the 0.99 (cosine) most similar 'spectral.events'.}

\item{top.expressing.override}{a named numeric vector -- default \code{NULL}; to override the automated selection of 'spectral.events', the name(s) of the supplied numeric vector should match to unique .fcs sample name(s). Example as follows: c("BUV615 (Cells)" = 20); for 'BUV615', only 20 top expressing events will be used to characterize/derive the final spectra (see example below).}

\item{print.timings}{logical -- default \code{FALSE}; if \code{TRUE}, a \link[tictoc:tic]{tictoc} will print to the console, detailing execution times.}

\item{...}{not defined; placeholder.}
}
\value{
A \link[data.table:data.table]{data.table} of normalized \verb{[0,1]} spectra; an additional \emph{generalized} autofluoresence signature is appended to the table.
}
\description{
Manufacturer::Cytometer::Software tested:
\itemize{
\item Cytek::Aurora(5L)::SpectroFlo 3.3.0
}

As a fully data-driven/automated process -- for best performance -- \code{spectracle} relies on adherence to established naming conventions for appropriate identification of the unstained control, parsing of keyword-value pairs into factored metadata, and accurate annotation of derived spectra.
}
\examples{
\dontrun{
raw.reference.controls <- fcs.files <- list.files(
system.file("extdata/expt1", package = "spectracle"),
full.names = TRUE
)

spectra <- spectracle(raw.reference.controls)

## name fix due to no marker '$PnS' names in these source files
## add empirically determined marker names
spectra[
i = N == '450',
j = c('N', 'S') := list('eFluor 450', 'TCRgd')
]
spectra[
i = N == 'BUV615',
j = c('N', 'S') := list('BUV615', 'Siglec F')
]

##
plot_trace(spectra[N == "eFluor 450"])

## this particular control is low-event count/rare
## impacted by AF -- the spectra is off...
plot_trace(spectra[N == "BUV615"])

## ...rerun with a small tweak
spectra <- spectracle(
raw.reference.controls[c(1,3)],
filter.top.expressing = c("BUV615 (Cells)")
)

## spectra is derived
plot_trace(spectra[N == "BUV615"])
}

\dontrun{
raw.reference.controls <- fcs.files <- list.files(
system.file("extdata/expt2", package = "spectracle"),
full.names = TRUE
)

spectra <- spectracle(raw.reference.controls)
spectra[, .(N, hash.md5)]

}
}
