% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectracle.R
\name{spectracle}
\alias{spectracle}
\title{Derive Spectra from Full Spectrum Cytometry Controls}
\usage{
spectracle(
  raw.reference.controls,
  filter.top.expressing = NULL,
  top.expressing.override = NULL,
  print.timings = FALSE,
  ...
)
}
\arguments{
\item{raw.reference.controls}{a character vector of full directory/path names to raw single-color control .fcs files; one of \link{list.dirs} or \link{list.files}.}

\item{filter.top.expressing}{a character vector -- default \code{NULL}; to modulate the automated selection of 'spectral.events', the supplied character vector should match to unique .fcs sample name(s). If defined, a conditional filtering will take place to return only the 0.99 (cosine) most similar 'spectral.events'.}

\item{top.expressing.override}{a named numeric vector -- default \code{NULL}; to override the automated selection of 'spectral.events', the name(s) of the supplied numeric vector should match to unique .fcs sample name(s). Example as follows: c("BUV615 (Cells)" = 20); for 'BUV615', only 20 top expressing events will be used to characterize/derive the final spectra (see example below).}

\item{print.timings}{logical -- default \code{FALSE}; if \code{TRUE}, a \link[tictoc:tic]{tictoc} will print to the console, detailing execution times.}

\item{...}{not defined; placeholder.}
}
\value{
A \link[data.table:data.table]{data.table} of normalized \verb{[0,1]} spectra; an additional \emph{generalized} autofluoresence signature is appended to the table.
}
\description{
Manufacturer::Cytometer::Software tested:
\itemize{
\item Cytek::Aurora(5L)::SpectroFlo 3.3.0
}

As a fully data-driven/automated process -- for best performance -- \code{spectracle} relies on adherence to established naming conventions for appropriate identification of the unstained control, parsing of keyword-value pairs into factored metadata, and accurate annotation of derived spectra.
}
\examples{

## due to large file sizes, the .fcs files used in this example were first
## fully processed -- spectra was derived and saved; load here

spectra.files <- list.files(
system.file("extdata/prepared_spectra", package = "spectracle"),
full.names = TRUE
)
names(spectra.files) <- sub(".rds", "", basename(spectra.files))

spectra.examples <- lapply(spectra.files, readRDS)

## expt1

spectra <- spectra.examples$spectra_expt1_nofilter
plot_trace(spectra[N == "eFluor 450"])

## this particular control is low-event count/rare
## impacted by AF -- the spectra is off...
plot_trace(spectra[N == "BUV615"])

## ...rerun with a small tweak using:
## `filter.top.expressing = c("BUV615 (Cells)")`
spectra <- spectra.examples$spectra_expt1_filtered

## spectra is derived
plot_trace(spectra[N == "BUV615"])


## expt2 -- beads and cells
spectra <- spectra <- spectra.examples$spectra_expt2

plot_trace(spectra)

spectra[, .(N, hash.md5)]

}
